Move video util functions
