+++ src/containers/TopicArticlePage/components/TopicArticleForm.jsx
@@ -29,6 +29,7 @@ import {
 import { FormikAlertModalWrapper, formClasses } from '../../FormikForm';
 import { toEditArticle } from '../../../util/routeHelpers';
 import { nullOrUndefined } from '../../../util/articleUtil';
+import { getYoutubeEmbedUrl, toHMS} from '../../../util/videoUtil';
 import validateFormik from '../../../components/formikValidationSchema';
 import TopicArticleAccordionPanels from './TopicArticleAccordionPanels';
 import HeaderWithLanguage from '../../../components/HeaderWithLanguage';
@@ -88,41 +89,6 @@ const getPublishedDate = (values, initialValues, preview = false) => {
   return undefined;
 };
 
-const toHMS = seconds => {
-  if (!seconds) return undefined;
-
-  const minute = Math.floor(seconds / 60) % 60;
-  const hour = Math.floor(minute / 60) % 60;
-  const second = seconds % 60;
-
-  const hours = hour > 0 ? hour + ':' : '';
-  const minutes = minute > 0 ? minute + ':' : '';
-  const secondos = second < 10 && second > 0 ? '0' + second : second;
-
-  return `${hours}${minutes}${secondos}`;
-};
-
-const calcSecondsFromHMS = hms => {
-  return hms
-    .split(':')
-    .reverse()
-    .map(a => parseInt(a, 10))
-    .filter(Number)
-    .reduce((acc, element, index) => acc + element * Math.pow(60, index));
-};
-
-const url = (url, start, stop) => {
-  if (!start && !stop) return url;
-  let youtube_embed_url = url.includes('embed')
-    ? `${url.split('?')[0]}?`
-    : `https://www.youtube.com/embed/${url.split('v=')[1]}?`;
-
-  if (start) youtube_embed_url += `&start=${calcSecondsFromHMS(start)}`;
-  if (stop) youtube_embed_url += `&end=${calcSecondsFromHMS(stop)}`;
-
-  return youtube_embed_url;
-};
-
 // TODO preview parameter does not work for topic articles. Used from PreviewDraftLightbox
 const getArticleFromSlate = ({
   values,
@@ -145,7 +111,7 @@ const getArticleFromSlate = ({
             values.visualElementAlt && values.visualElementAlt.length > 0
               ? values.visualElementAlt
               : undefined,
-          url: url(
+          url: getYoutubeEmbedUrl(
             values.visualElement.url,
             values.visualElementStart,
             values.visualElementStop,
